+++ src/test/java/org/overrun/glutest/Tesselator3Test.java
@@ -95,7 +95,7 @@ public class Tesselator3Test extends Game {
 
     @Override
     public void render() {
-        float delta = timer.getDelta();
+        float delta = (float) timer.getDelta();
         float tx = xo + (x - xo) * delta;
         float ty = yo + (y - yo) * delta;
         float tz = zo + (z - zo) * delta;
